requirements.txt
flask
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
openai
pypdf
gunicorn